requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
pytest
openai==0.28
apscheduler
cachetools
pytest